Support ansible 2.4 with vendored library
This is effectively a proof of concept for a way to backport a module
from newer versions of ansible to work with a version of ansible that
did not have that module. In this case, rhsm_repository does not exist
in ansible 2.4. If (and only if) the installed version of ansible is
less than 2.5, this role will attempt to use the vendored version of the
task. Otherwise (ansible 2.5+), it will use the ansible builtin version
of the task.