Add clarification of role use to README
In the process of actually using this role, I was reminded that it
exists mainly to address a bit of an edge case, which is to be able to
add an address to an interface after that interface has already been
configured (e.g. after installing a system that receives addresses via
DHCP, this role can be used to add an additional static address to that
interface).