Move test scenarios upstream
This does a few things:
- With oasis-pipeline on github, all the tests can be on github now,
too. The only thing I see that needed to be redacted is the subscription
information, which now comes in via jankies env vars. The majority of
this change is the addition of a group_vars file to load the
subscription information from the env, and then updating the scenarios
to use those group_vars.
- The "downstream_openstack" scenario has entirely replaced the
"openstack" scenario previously in this repo.
- The Jenkinsfile itself now lives in this repo

If the rhsm group_vars approach deemed is good, I'll be documenting
that so it can easily be adapted to other roles that need subscriptions
for testing.