Remove pool comparison from output test
Due to changes in how RHSM vars are loaded, getting at the list of
expected pool IDs from outside of ansible is a little less reliable.
This test is modified just to check that the subscribed pool IDs are
correctly added to the output yaml.

I also added some entries to the gitignore to ignore test-generated
artifacts.