Add support for OpenStack VM volumes
In the molecule.yml file, a top level field "volumes" can be created
that is an array of os_volume arguments, including name and size. These
resources will be created and destroyed just as the VMs are. Also, this
allows the enhancement of the creation of a VM with the "volumes"
field inside its definition dict.