Various CI updates and improvements
In documenting ways to best test roles that require subscriptions, and
in light of changes either recently applied to or soon coming to the
molecule_openstack_ci role, I made a few changes to test the docs that
I'm writing.

This commit attempts to clean up the rhsm role to more properly work
with the new envvar-based method for getting at the subscription
information, removes the unused docker scenario, and fixes up the
openstack provisioning playbooks to better match how the
molecule_openstack_ci does things.