Remove subscription-manager install task
On a system that needs to be registered, subscription-manager should
already be installed. If subscription-manager is not installed (this
should never happen), it would then need to be installed, which would
require a registered system, which would have required subscription-
manager to be installed.

tl;dr you can't install subscription-manager without subscription-
manager, so...we can safely assume it's installed.