Update GitHub Actions
Running a separate lint step is dumb. Every role lints itself
Failing fast is dumb. We want as many of the role executions to run to
completion as we can get so we have maximum feedback
libvirt_rhel_vm requires a VM, doesn't seem to run upstream until we can
get network_bridge working on Ubuntu